backend/requirements.txt
# Django and Django REST Framework
Django==5.2.3
djangorestframework==3.15.2
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter (for production)

# Authentication and Authorization
djangorestframework-simplejwt==5.3.0

# File handling and validation
Pillow==10.1.0
python-magic==0.4.27

# API documentation
drf-yasg==1.21.7

# Environment variables
python-decouple==3.8

# AI and ML libraries
google-generativeai>=0.3.0
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0

# OCR and document processing
pytesseract==0.3.10
PyPDF2==3.0.1
python-docx==1.1.0

# HTTP requests
requests==2.31.0
httpx==0.25.2

# Data processing
pandas==2.1.4
numpy==1.25.2

# Search and indexing
elasticsearch==8.11.1

# Caching
redis==5.0.1
django-redis==5.4.0

# Task queue
celery==5.3.4

# Monitoring and logging
sentry-sdk==1.38.0

# Development tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Linting and formatting
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30

# Web scraping and search
beautifulsoup4==4.12.2
selenium==4.15.2
serpapi==0.1.5
